requirements.txt
certifi==2018.11.29
chardet==3.0.4
click==6.7
colorama==0.4.1
Django==2.1.7
django-cors-headers==2.4.1
django-filter==2.1.0
django-hashid-field==2.1.5
djangorestframework==3.9.1
hashids==1.2.0
idna==2.8
pyconfiglib==1.3.3
Pygments==2.3.1
pypandoc==1.4
pytz==2018.9
requests==2.21.0
superprompt==1.2.0
urllib3==1.24.1
